chore(word counts) - Moved the word counts on etree module
In order to cut the dependence between archive and io I moved the
get_word_count function on etree module.
